+++ pkgs/google_generative_ai/lib/google_generative_ai.dart
@@ -65,11 +65,15 @@ export 'src/api.dart'
 export 'src/chat.dart' show ChatSession, StartChatExtension;
 export 'src/content.dart'
     show
+        CodeExecutionResult,
         Content,
         DataPart,
+        ExecutableCode,
         FilePart,
         FunctionCall,
         FunctionResponse,
+        Language,
+        Outcome,
         Part,
         TextPart;
 export 'src/error.dart'
@@ -80,6 +84,7 @@ export 'src/error.dart'
         UnsupportedUserLocation;
 export 'src/function_calling.dart'
     show
+        CodeExecution,
         FunctionCallingConfig,
         FunctionCallingMode,
         FunctionDeclaration,